Move RefItem file
